requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
httpx==0.28.1
openai==1.59.7
python-dotenv==1.0.1
pydantic==2.10.5
aiosqlite==0.21.0
beautifulsoup4==4.12.3
curl_cffi==0.7.4
playwright==1.49.1
